+++ RELEASE-NOTES.md
@@ -9,6 +9,7 @@ Development version
 * Dropped support for PHP older than 5.5
 * Fixed missing marker cluster images for Google Maps (by Peter Grassberger)
 * Fixed duplicate markers in OpenLayers maps (by Peter Grassberger)
+* Added cluster properties for Google Maps (by Peter Grassberger)
 
 ## Maps 3.5